pyproject.toml
[project]
name = "nlvm"
version = "0.1.0"
description = "An natural language runtime and python binding toolkit"
authors = [
    {name = "倪旌哲",email = "1448976235@qq.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic>=2.5,<3.0",
    "simplellmfunc (==0.6.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.6,<1.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[dependency-groups]
dev = [
    "pytest (>=8.0,<9.0)",
    "pytest-asyncio (>=0.23,<0.24)",
    "ruff (>=0.6,<0.7)"
]
